+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 [Full changelog](https://github.com/mozilla/glean/compare/v34.0.0...v34.1.0)
 
+* General
+  * A new metric `glean.validation.pings_submitted` tracks the number of pings sent. It is included in both the `metrics` and `baseline` pings.
 * iOS
   * The metric `glean.validation.foreground_count` is now sent in the metrics ping ([#1472](https://github.com/mozilla/glean/pull/1472)).
   * BUGFIX: baseline pings with reason `dirty_startup` are no longer sent if Glean did not full initialize in the previous run ([#1476](https://github.com/mozilla/glean/pull/1476)).